# specs/openapi/bundled.yaml
openapi: "3.1.0"
info:
  title: Consensus Agent Runtime
  description: |
    Database-native cognitive architecture for AI agents.

    All state lives in the database. The API writes to the same tables the agent reads from.
    There is no separate API state — the database is the single source of truth.

    ## Key Concepts
    - **Sessions** — Every agent instance is a session row. RLS isolation, billing, lifecycle.
    - **Memory Events** — Immutable append-only ledger. Display state managed via display_modes.
    - **Transactions** — Atomic cognition: every thought is a SQL transaction.
    - **Tools & Skills** — Progressive disclosure: metadata always visible, full instructions on demand.
    - **HITL Approvals** — Human-in-the-loop: pause for review before destructive actions.

    ## Authentication
    All endpoints except `/api/v1/health` require a Bearer API key:
    - cs_ak_* : admin — full CRUD
    - cs_sk_* : session — scoped to one session
    - cs_rk_* : readonly — SELECT only
    - cs_wk_* : webhook — event ingestion only

    Keys are SHA-256 hashed before storage. The full key is only returned at creation time.
  version: "0.1.0"
  contact:
    name: Consensus
    url: https://github.com/wojons/consensus
servers:
  - url: http://localhost:8090
    description: Local development (SQLite or local Postgres)
  - url: https://{host}
    description: Remote deployment (any Postgres provider)
    variables:
      host:
        default: localhost:8090
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/sessions:
    post:
      tags: [Sessions]
      summary: Create a new agent session
      operationId: createSession
      security: [{BearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Sessions]
      summary: List sessions
      operationId: listSessions
      security: [{BearerAuth: []}]
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filter by comma-separated status values.
      responses:
        "200":
          description: List of sessions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key prefixed with cs_ (e.g., cs_sk_a1b2c3d4...).
  schemas:
...
    CreateSessionRequest:
      type: object
      required: [agent_name, goal]
      properties:
        agent_name:
          type: string
        goal:
          type: string
        model_id:
          type: string
        context_budget:
          type: integer
        hitl_config:
          type: object
          additionalProperties: true
    CreateSessionResponse:
      type: object
      required: [id, status, api_key, created_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        api_key:
          type: string
        created_at:
          type: string
          format: date-time
    Session:
      type: object
      required: [id, agent_name, model_id, status, context_budget, iteration, created_at]
      properties:
        id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        agent_name:
          type: string
        model_id:
          type: string
        status:
          type: string
          enum: [booting, idle, thinking, planning, tool_exec, executing, waiting_sub, paused, completed, failed]
        goal:
          type: string
          nullable: true
        context_budget:
          type: integer
        tokens_used_in:
          type: integer
          format: int64
        tokens_used_out:
          type: integer
          format: int64
        iteration:
          type: integer
          format: int64
        heartbeat_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
...
  responses:
...
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/responses/ErrorResponse/content/application~1json/schema"
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/responses/ErrorResponse/content/application~1json/schema"
    Forbidden:
      description: API key scope does not permit this action.
      content:
        application/json:
          schema:
            $ref: "#/components/responses/ErrorResponse/content/application~1json/schema"
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/responses/ErrorResponse/content/application~1json/schema"